Switch to packed result storage and more
Remove the KP_CAP_CH_STATUS_DISABLED, and instead check the channel
configuration and capture direction to figure out if the channel is
disabled, or not. This allows packed result storage, as uncaptured
channels don't need to be stored.

Switch to supplying a configuration structure, instead of separate
parameters to kp_cap_start(). This way the call is not so messy and we
can easily copy capture settings.

Rename `kp_cap_dir` to `kp_cap_dirs` emphasising that it's a set
(bitmap) that can contain multiple directions. Add another enumeration
for all possible non-empty direction sets: kp_cap_ne_dirs, and use it to
index summarized data after measuring.

Switch to packed result storage, so that we don't have memory wasted for
disabled channels/directions.

Yes, this should've been multiple commits, but I didn't have time to
split it, so holding my fingers crossed 🤞